+++ lib/gimei.rb
@@ -41,6 +41,18 @@ class Gimei
     def unique(max_retries = 10_000)
       @unique ||= UniqueGenerator.new(max_retries)
     end
+
+    def config
+      @_config ||= Config.new
+    end
+  end
+
+  class Config
+    attr_accessor :rng
+
+    def initialize
+      @rng = Random.new
+    end
   end
 
   def initialize(gender = nil)