+++ lib/gimei/name.rb
@@ -22,7 +22,7 @@ class Gimei::Name
   end
 
   def initialize(gender = nil)
-    @gender = gender || Gimei::GENDERS.sample
+    @gender = gender || Gimei::GENDERS.sample(random: Gimei.config.rng)
     @first = First.new @gender
     @last = Last.new
   end
@@ -68,8 +68,8 @@ class Gimei::Name
     def_delegators :@name, :kanji, :hiragana, :katakana, :to_s, :romaji
 
     def initialize(gender = nil)
-      @gender = gender || Gimei::GENDERS.sample
-      @name = NameWord.new(Gimei::NAMES['first_name'][@gender.to_s].sample)
+      @gender = gender || Gimei::GENDERS.sample(random: Gimei.config.rng)
+      @name = NameWord.new(Gimei::NAMES['first_name'][@gender.to_s].sample(random: Gimei.config.rng))
     end
 
     def male?
@@ -86,7 +86,7 @@ class Gimei::Name
     def_delegators :@name, :kanji, :hiragana, :katakana, :to_s, :romaji
 
     def initialize
-      @name = NameWord.new(Gimei::NAMES['last_name'].sample)
+      @name = NameWord.new(Gimei::NAMES['last_name'].sample(random: Gimei.config.rng))
     end
   end